requirements.txt
ragas==0.1.21
deepeval==0.21.73
langchain==0.2.16
langchain-openai==0.1.23
langchain-google-vertexai==1.0.10
openai==1.45.0
google-cloud-aiplatform==1.65.0
pandas==2.2.2
numpy==1.26.4
pytest==8.3.3
pytest-asyncio==0.23.8
allure-pytest==2.13.5
python-dotenv==1.0.1
pydantic==2.8.2
rich==13.8.1
jinja2==3.1.4
httpx==0.27.2
